+++ src/gamestates/cOptionsState.h
@@ -3,7 +3,7 @@
 #include "cGameState.h"
 #include "controls/cKeyboardEvent.h"
 #include "drawers/cTextDrawer.h"
-#include "gui/cGuiWindow.h"
+#include "gui/GuiWindow.h"
 #include "sMouseEvent.h"
 
 struct SDL_Surface;
