+++ src/gui/GuiObject.h
