+++ src/gui/actions/cGuiActionShowOptions.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gui/cGuiAction.h"
+#include "gui/GuiAction.h"
 
 class cGuiActionShowOptions : public cGuiAction {
 public:
