--- src/gui/cGuiAction.cpp
@@ -1,1 +1,0 @@
-#include "cGuiAction.h"
